# Use atomic.Value rather than unsafe.Pointer