@julz 2021-03-05
yeah, will update the other PR since it hasn't landed yet (waiting for the .github change to propagate to bump golanglint-ci version)