@markusthoemmes 2021-03-05
Huh? The other thing is importas, right? That doesn't have anything to do with this and the atomic linter stuff landed :thinking:. So I'd think you can drop the exception in this PR.